requirements.txt
pandas     # data manipulation
openpyxl   # Excel (.xlsx) reading
pyyaml     # local path config
